Show warnings for FIXMEs.
This was tested out in CNTRAL Mobile™ first and it was extremely useful for catching and preventing FIXMEs from entering into the master branch.

You can use FIXMEs while developing to quickly prototype something or whatever. But they should never be forgotten and ignored when merging into master so throw warnings so it's caught by PR checks, etc.